### 2020-10-28 #2787 @PatrickRWright
Suggestions for reviewers:
masalmon, Tgerke, acolum, pacoramon, Jminnier, sachsmc, rrrlw

Really anyone in [the list](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822) which a background in R and medicine/health/epidemiology.
