### 2020-10-28 #2787 @PatrickRWright
Hey, @aghaynes and @markomi this is the thread for the secuTrialR submission to JOSS.
