### 2020-10-29 #2787 @csoneson
👋 @PatrickRWright - I will handle your submission, and start by inviting suitable reviewers. In the meanwhile, could you please add a 'Statement of Need' section to your paper (see e.g. [here](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain))? Also, during a quick read-through I noticed that there was one place in the manuscript where `CDMS` was misspelled as `CMDS`.
