### 2020-10-31 #2787 @pacoramon
Dear @csoneson, I would be interested although being my first time reviewing for JOSS (and in this format), I am concerned I will need a bit of help to manage.
Would that be 'ok'?  Could you guide me to any documentation or guideline from JOSS I have to be aware off?
