### 2020-10-31 #2787 @csoneson
Thank you @pacoramon - that's not a problem at all. JOSS guidelines for reviewers are available [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

Briefly, the process is as follows: I will assign you as a reviewer in this thread, but the 'actual' review will start only once we have one additional reviewer. At that time, I will open the REVIEW GitHub issue, in which you will find a personal checklist with the aspects that we ask our reviewers to evaluate (the checklist is also available in the documentation linked above). You check off the items that you are satisfied with, and for the others, provide comments to the authors either in the REVIEW issue itself, or by opening separate issues in the software repository itself. If you have questions at any time, feel free to ping me.
