### 2020-10-31 #2787 @pacoramon
Thank you @csoneson for your kind explanation. Count me in, then. I will review the JOSS guidelines while another reviewer is enlisted.
Best!
