### 2020-11-03 f00ea58
stop tcltk and igraph autoimport
