### 2020-11-04 #2787 @csoneson
👋 @tgerke, @sachsmc - would one of you be willing to review this submission for JOSS?

> secuTrialR: Seamless interaction with clinical trial databases in R
