### 2020-11-04 #2816 @whedon
[REVIEW]: secuTrialR: Seamless interaction with clinical trial databases in R
**Submitting author:** @PatrickRWright (<a href="http://orcid.org/0000-0002-1153-0846">Patrick R. Wright</a>)
**Repository:** <a href="https://github.com/SwissClinicalTrialOrganisation/secuTrialR" target ="_blank">https://github.com/SwissClinicalTrialOrganisation/secuTrialR</a>
**Version:** 1.0.6
