### 2020-11-06 SwissClinicalTrialOrganisation/secuTrialR#225 @sachsmc
Add URL and BugReports fields to DESCRIPTION
This is for the JOSS review, checklist item: "Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"

**Describe the bug**
Users that find your package on CRAN may have a hard time finding the relevant guidelines to report a bug or to browse the source code (available only at the bottom of the README currently). This can be solved by adding the following to the DESCRIPTION file

    URL: https://github.com/SwissClinicalTrialOrganisation/secuTrialR
    BugReports: https://github.com/SwissClinicalTrialOrganisation/secuTrialR/issues
