### 2020-11-06 SwissClinicalTrialOrganisation/secuTrialR#226 @sachsmc
JOSS Software paper issues
A few notes on the JOSS software paper:

1. For the checklist item: State of the field: Do the authors describe how this software compares to other commonly-used packages?

This could be improved. For example, what are the existing approaches to interact with data from secuTrial? How does the functionality in secuTrialR compare with that of R interfaces to competing clinical data management systems such as redcapR and ox (openclinica)?

2. Issues in the references:

- Kelion 2020 -- Microsoft and COVID-19 should be capitalized
- Kuchinke -- Europe and ECRIN
- Merlot -- Url runs off page
- Rstudio and Wickham -- R should be capitalized
