### 2020-11-06 #2816 @sachsmc
Hi @PatrickRWright ,

I noted a few small issues with the documentation and the software paper in issues 225 and 226 on your repo:

https://github.com/SwissClinicalTrialOrganisation/secuTrialR/issues/225
https://github.com/SwissClinicalTrialOrganisation/secuTrialR/issues/226

Otherwise the package looks and works great!

Best wishes,

Michael
