### 2020-11-09 SwissClinicalTrialOrganisation/secuTrialR#226 @PatrickRWright
@aghaynes off the top of your head could you write something comparing `secuTrialR` to `REDCapR`?
