### 2020-11-09 SwissClinicalTrialOrganisation/secuTrialR#225 @PatrickRWright
Thanks @sachsmc for the useful suggestion.
