### 2020-11-09 SwissClinicalTrialOrganisation/secuTrialR#226 @aghaynes
Yes, I can do that, maybe this afternoon or tomorrow... OK?
