### 2020-11-09 SwissClinicalTrialOrganisation/secuTrialR#226 @PatrickRWright
Super thanks Alan. I don't think there's a rush.

While you're editing the paper, can you also add the two reviewers and the editor to the acknowledgements list?
