+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
   secuTrialR-package-vignette: secuTrialR-package-vignette.html
-last_built: 2020-11-03T11:45Z
+last_built: 2020-11-11T07:31Z
 
