### 2020-11-11 bcade51
modifications from review
