### 2020-11-11 7fa715e
additional fixes to references
