+++ paper/paper.bib
@@ -22,7 +22,7 @@
   year = {2010},
   month = {07},
   pages = {79},
-  title = {Heterogeneity prevails: The state of clinical trial data management in Europe - results of a survey of ECRIN centres},
+  title = {Heterogeneity prevails: The state of clinical trial data management in {Europe} - results of a survey of {ECRIN} centres},
   volume = {11},
   journal = {Trials},
   doi = {10.1186/1745-6215-11-79}
@@ -52,7 +52,7 @@
   author = {Kelion, L},
   year = {2020},
   month = {10},
-  title = {Excel: Why using Microsoft's tool caused Covid-19 results to be lost},
+  title = {Excel: Why using {Microsoft's} tool caused {COVID-19} results to be lost},
   howpublished = {\url{https://www.bbc.com/news/technology-54423988}},
   note = {Accessed: 2020-10-16}
 }
@@ -78,7 +78,7 @@
 }
 
 @Manual{devtools_cit,
-    title = {devtools: Tools to Make Developing R Packages Easier},
+    title = {{devtools}: Tools to Make Developing {R} Packages Easier},
     author = {Hadley Wickham and Jim Hester and Winston Chang},
     year = {2019},
     note = {R package version 2.2.1},
@@ -86,7 +86,7 @@
 }
 
 @Manual{rstudio_cit,
-    title = {RStudio: Integrated Development Environment for R},
+    title = {RStudio: Integrated Development Environment for {R}},
     author = {{RStudio Team}},
     organization = {RStudio, Inc.},
     address = {Boston, MA},
