+++ paper/paper.md
@@ -119,7 +119,7 @@ Thomas Fabbro, Inessa Kraft, Arnaud Künzi, Daniel Lengwiler, Armando Lenz, Henr
 Marie Roumet, Constantin Sluka, Suvitha Subramaniam, Miriam Wegmann, Laura Werlen and Thomas Zumbrunn for ideas,
 testing and constructive feedback on the `secuTrialR` package. We also thank [Michael Sachs](https://github.com/sachsmc) 
 and [pacoramon](https://github.com/pacoramon) for kindly reviewing this manuscript and the R package and making additional 
-recommendations.
+recommendations, and [Charlotte Soneson](https://github.com/csoneson) for acting as editor.
 Furthermore, the authors thank the State Secretariat of Education, Research and Innovation and the Swiss National
 Science Foundation for the funding of this project and the Swiss Clinical Trial Organisation for its ongoing support.
 
