### 2020-11-11 #2816 @csoneson
> We are, however, not sure how to solve the issue with one of the links in the references moving out of the page. Any idea?

Good question. I'm not sure why that link is not wrapped like the others. Perhaps someone from the @openjournals/dev team has an idea.
