### 2020-11-11 #2816 @sachsmc
@csoneson  One way to solve it requires modifying the latex template (not sure how onerous that would be). If you add the code

    \usepackage[hyphens]{url}

to the header before the hyperref package is loaded, it will break the url at hyphens. By default, it will only break at slashes.

https://tex.stackexchange.com/questions/175399/line-breaking-of-urls-at
