### 2020-11-16 #2816 @xuanxu
> We are, however, not sure how to solve the issue with one of the links in the references moving out of the page. Any idea?

A quick fix would be to use the shorter-form url for the same article:
https://www.spiegel.de/wissenschaft/a-13bd06d7-22a1-4b3d-af23-ff43e5e8abd6
