+++ paper/paper.bib
@@ -44,7 +44,7 @@
   year = {2020},
   month = {03},
   title = {Statistikprobleme beim Coronavirus - Die große Meldelücke},
-  howpublished = {\url{https://www.spiegel.de/wissenschaft/coronavirus-wie-belastbar-sind-die-rki-daten-a-13bd06d7-22a1-4b3d-af23-ff43e5e8abd6}},
+  howpublished = {\url{https://www.spiegel.de/a-13bd06d7-22a1-4b3d-af23-ff43e5e8abd6}},
   note = {Accessed: 2020-04-06}
 }
 
