+++ paper/paper.bib
@@ -112,21 +112,21 @@
     pages = "103208",
     year = "2019",
     issn = "1532-0464",
-    doi = "https://doi.org/10.1016/j.jbi.2019.103208",
+    doi = {10.1016/j.jbi.2019.103208},
     url = "http://www.sciencedirect.com/science/article/pii/S1532046419301261",
     author = "Paul A. Harris and Robert Taylor and Brenda L. Minor and Veida Elliott and Michelle Fernandez and Lindsay O'Neal and Laura McLeod and Giovanni Delacqua and Francesco Delacqua and Jacqueline Kirby and Stephany N. Duda",
     keywords = "Medical informatics, Electronic data capture, Clinical research, Translational research"
 }
 
 @Article{Harris2009,
-    title = "Research electronic data capture (REDCap)—A metadata-driven methodology and workflow process for providing     translational research informatics support",
+    title = "Research electronic data capture (REDCap)—A metadata-driven methodology and workflow process for providing translational research informatics support",
     journal = "Journal of Biomedical Informatics",
     volume = "42",
     number = "2",
     pages = "377 - 381",
     year = "2009",
     issn = "1532-0464",
-    doi = "https://doi.org/10.1016/j.jbi.2008.08.010",
+    doi = {10.1016/j.jbi.2008.08.010},
     url = "http://www.sciencedirect.com/science/article/pii/S1532046408001226",
     author = "Paul A. Harris and Robert Taylor and Robert Thielke and Jonathon Payne and Nathaniel Gonzalez and Jose G. Conde",
     keywords = "Medical informatics, Electronic data capture, Clinical research, Translational research"
