### 2020-11-18 #2816 @csoneson
Thanks @PatrickRWright, looks good. I had a look through the submission, and I just noted one other thing: I think there is at least one place where `secuTrialdata` should be in fixed-width font, could you double-check this?

The next steps are as follows:
- [ ] Make a tagged release of the current state of your software, and list the version tag of the archived version here.
- [ ] Check that the archival deposit has the correct metadata. This includes the title (should match the paper title) and author list (should match the paper author list). You may also add the authors' ORCID.
