### 2020-11-18 #2816 @PatrickRWright
Thank you @csoneson
Ther version tag would be: 1.0.6
Zenodo link: https://zenodo.org/record/4278979#.X7TdUq4o-EL
DOI: seems to be the one I entered for this publication. i.e. 10.21105/joss.02816
