### 2020-11-18 #2816 @csoneson
Ah, I see. The Zenodo archive should have its own DOI, something like `10.5281/zenodo.XXXXXX` (i.e., not the one of the JOSS paper).
