### 2020-11-18 #2816 @PatrickRWright
Hm... I'm a bit confused now. since the DOI tooltip for zenodo said:
```
Optional. Did your publisher already assign a DOI to your upload? If not, leave
the field empty and we will register a new DOI for you. A DOI allows others to
easily and unambiguously cite your upload. Please note that it is NOT possible
to edit a Zenodo DOI once it has been registered by us, while it is always
possible to edit a custom DOI.
```
For me this indicated to use a preexisting DOI (i.e. JOSS) if available. I tried to remove the JOSS DOI now, but saving on zenodo doesn't actually remove it. Any suggestions?
