### 2020-11-18 #2816 @danielskatz
Zenodo will issue its own DOI for the deposit (which you will see once it is published "Cite as" on the lower right).  In addition, you can put other DOIs into the page where you submit to zenodo, such as references and links. These DOIs are stored in the metadata of the published article under the published DOI.

When you manually upload to Zenodo, there's a place for a DOI, which says "Optional. Did your publisher already assign a DOI to your upload? If not, leave the field empty and we will register a new DOI for you. A DOI allows others to easily and unambiguously cite your upload. Please note that it is NOT possible to edit a Zenodo DOI once it has been registered by us, while it is always possible to edit a custom DOI."  In the case of JOSS, we do not assign a DOI to the software upload, so you should leave this empty.  We will assign a DOI to the paper, which will refer to the DOI of the software upload (which is why we need you to tell us what it at this stage of the acceptance process).
