### 2020-11-18 #2816 @csoneson
Thanks @danielskatz for the more extensive explanation.

@PatrickRWright - I think you may need to email Zenodo to resolve this and replace the JOSS DOI with a Zenodo-assigned one.
