### 2020-11-19 #2816 @PatrickRWright
@csoneson Zenodo suggested to reupload the package leaving the DOI field empty. They will then delete the old submission.

I have done this: https://zenodo.org/record/4280767#.X7Zbqa4o_MU
DOI: http://doi.org/10.5281/zenodo.4280767
Version: 1.0.6
