### 2020-11-20 #2816 @PatrickRWright
Should I run this `@whedon accept deposit=true` or is the something the associate editor needs to do?

Also I saw that the XML omits the middle names. i.e. `Alan Haynes` should be `Alan G. Haynes` and `Patrick Wright` should be `Patrick R. Wright`. Is that possible?
