### 2020-11-20 #2816 @danielskatz
I'll take over from here, with a final proofreading and then the acceptance.  I'll also look at the name issue in the XML - thanks for catching it as a potential issue.
