### 2020-11-20 #2816 @danielskatz
Please merge the changes in https://github.com/SwissClinicalTrialOrganisation/secuTrialR/pull/229 or let me know what you don't agree with.
