### 2020-11-20 #2816 @danielskatz
@PatrickRWright - It turns out that Crossref only wants the given name (e.g. Alan). We omit the middle initial deliberately. But since we add the ORCID, the authors should still be uniquely identified.
